requirements.txt
langchain==1.0.1
python-dotenv==1.2.1
langchain-groq==1.0.0
numpy==2.3.4
pandas==2.3.3
quantreo==0.1.0